Split out 2 classes in InterfaceFactory files
