+++ qt/widgets/common/CMakeLists.txt
@@ -25,7 +25,6 @@ set(QT5_SRC_FILES
     # todo: move this to the instrument view library when the slice
     # viewer library is removed
     src/InputController.cpp
-    src/InterfaceFactory.cpp
     src/InterfaceManager.cpp
     src/LineEditWithClear.cpp
     src/ListPropertyWidget.cpp
@@ -62,6 +61,7 @@ set(QT5_SRC_FILES
     src/UserFunctionDialog.cpp
     src/UserInputValidator.cpp
     src/UserSubWindow.cpp
+    UserSubWindowFactory.cpp
     src/VatesViewerInterface.cpp
     src/WidgetScrollbarDecorator.cpp
     src/WorkspaceIcons.cpp
@@ -194,6 +194,7 @@ set(QT5_INC_FILES
     ${QT5_MOC_FILES}
     inc/MantidQtWidgets/Common/AlgorithmInputHistory.h
     inc/MantidQtWidgets/Common/AlgorithmRunner.h
+    inc/MantidQtWidgets/Common/AlgorithmDialogFactory.h
     inc/MantidQtWidgets/Common/AlternateCSPythonLexer.h
     inc/MantidQtWidgets/Common/BatchAlgorithmRunner.h
     inc/MantidQtWidgets/Common/DropEventHelper.h
@@ -201,7 +202,6 @@ set(QT5_INC_FILES
     inc/MantidQtWidgets/Common/FlowLayout.h
     inc/MantidQtWidgets/Common/HelpWindow.h
     inc/MantidQtWidgets/Common/IFunctionBrowser.h
-    inc/MantidQtWidgets/Common/InterfaceFactory.h
     inc/MantidQtWidgets/Common/MantidDesktopServices.h
     inc/MantidQtWidgets/Common/MantidTreeWidgetItem.h
     inc/MantidQtWidgets/Common/MultifitSetupDialog.h
@@ -262,7 +262,6 @@ set(SRC_FILES
     src/HelpWindow.cpp
     src/FlowLayout.cpp
     src/Hint.cpp
-    src/InterfaceFactory.cpp
     src/InterfaceManager.cpp
     src/ListPropertyWidget.cpp
     src/ManageUserDirectories.cpp
@@ -293,6 +292,7 @@ set(SRC_FILES
     src/TSVSerialiser.cpp
     src/UserInputValidator.cpp
     src/UserSubWindow.cpp
+    src/UserSubWindowFactory.cpp
     src/VatesViewerInterface.cpp
     src/WidgetScrollbarDecorator.cpp
     src/WindowIcons.cpp
@@ -520,6 +520,7 @@ set(INC_FILES
   inc/MantidQtWidgets/Common/AlgorithmInputHistory.h
   inc/MantidQtWidgets/Common/AlgorithmRunner.h
   inc/MantidQtWidgets/Common/AlternateCSPythonLexer.h
+  inc/MantidQtWidgets/Common/AlgorithmDialogFactory.h
   inc/MantidQtWidgets/Common/BatchAlgorithmRunner.h
   inc/MantidQtWidgets/Common/Configurable.h
   inc/MantidQtWidgets/Common/DllOption.h
@@ -530,7 +531,6 @@ set(INC_FILES
   inc/MantidQtWidgets/Common/DistributionOptions.h
   inc/MantidQtWidgets/Common/HelpWindow.h
   inc/MantidQtWidgets/Common/Hint.h
-  inc/MantidQtWidgets/Common/InterfaceFactory.h
   inc/MantidQtWidgets/Common/InterfaceManager.h
   inc/MantidQtWidgets/Common/IProjectSerialisable.h
   inc/MantidQtWidgets/Common/MantidDesktopServices.h
@@ -547,6 +547,7 @@ set(INC_FILES
   inc/MantidQtWidgets/Common/ScriptRepositoryView.h
   inc/MantidQtWidgets/Common/SelectionNotificationService.h
   inc/MantidQtWidgets/Common/SignalBlocker.h
+  inc/MantidQtWidgets/Common/UserSubWindowFactory.h
   inc/MantidQtWidgets/Common/TSVSerialiser.h
   inc/MantidQtWidgets/Common/UserInputValidator.h
   inc/MantidQtWidgets/Common/WidgetScrollbarDecorator.h
