+++ qt/widgets/common/inc/MantidQtWidgets/Common/AlgorithmDialog.h
@@ -24,8 +24,8 @@
 //----------------------------------
 // Includes
 //----------------------------------
+#include "AlgorithmDialogFactoryImpl.h"
 #include "DllOption.h"
-#include "InterfaceFactory.h"
 
 // Could have forward declared this but it makes it easier to use from
 // inheriting classes if it is included here
