+++ qt/widgets/common/inc/MantidQtWidgets/Common/UserSubWindow.h
@@ -24,8 +24,8 @@
 // Includes
 //----------------------------------
 #include "DllOption.h"
-#include "InterfaceFactory.h"
 #include "PythonRunner.h"
+#include "UserSubWindowFactoryImpl.h"
 
 #include <QLabel>
 #include <QMainWindow>
