+++ qt/widgets/common/src/InterfaceManager.cpp
@@ -9,12 +9,13 @@
 //----------------------------------
 #include "MantidQtWidgets/Common/InterfaceManager.h"
 #include "MantidQtWidgets/Common/AlgorithmDialog.h"
+#include "MantidQtWidgets/Common/AlgorithmDialogFactory.h"
 #include "MantidQtWidgets/Common/GenericDialog.h"
-#include "MantidQtWidgets/Common/InterfaceFactory.h"
 #include "MantidQtWidgets/Common/MantidDesktopServices.h"
 #include "MantidQtWidgets/Common/MantidHelpWindow.h"
 #include "MantidQtWidgets/Common/PluginLibraries.h"
 #include "MantidQtWidgets/Common/UserSubWindow.h"
+#include "MantidQtWidgets/Common/UserSubWindowFactory.h"
 #include "MantidQtWidgets/Common/VatesViewerInterface.h"
 
 #include "MantidAPI/AlgorithmManager.h"
@@ -175,14 +176,7 @@ UserSubWindow *InterfaceManager::createSubWindow(const QString &interface_name,
  * refer to UserSubWindow classes
  */
 QStringList InterfaceManager::getUserSubWindowKeys() const {
-  QStringList key_list;
-  std::vector<std::string> keys = UserSubWindowFactory::Instance().getKeys();
-  std::vector<std::string>::const_iterator iend = keys.end();
-  for (std::vector<std::string>::const_iterator itr = keys.begin(); itr != iend;
-       ++itr) {
-    key_list.append(QString::fromStdString(*itr));
-  }
-  return key_list;
+  return UserSubWindowFactory::Instance().getUserSubWindowKeys();
 }
 
 //----------------------------------
