+++ qt/widgets/common/src/UserSubWindowFactory.cpp
@@ -1,13 +1,13 @@
 // Mantid Repository : https://github.com/mantidproject/mantid
 //
-// Copyright &copy; 2018 ISIS Rutherford Appleton Laboratory UKRI,
+// Copyright &copy; 2019 ISIS Rutherford Appleton Laboratory UKRI,
 //     NScD Oak Ridge National Laboratory, European Spallation Source
 //     & Institut Laue - Langevin
 // SPDX - License - Identifier: GPL - 3.0 +
 //-----------------------------------------------
 // Includes
 //-----------------------------------------------
-#include "MantidQtWidgets/Common/InterfaceFactory.h"
+#include "MantidQtWidgets/Common/UserSubWindowFactoryImpl.h"
 #include "MantidKernel/Logger.h"
 #include "MantidQtWidgets/Common/UserSubWindow.h"
 #include <sstream>
@@ -112,4 +112,59 @@ UserSubWindowFactoryImpl::createFromAlias(const std::string &name) const {
   } else {
     return nullptr;
   }
+}
+
+template <typename TYPE> void UserSubWindowFactoryImpl::subscribe() {
+  std::string realName = TYPE::name();
+  Mantid::Kernel::DynamicFactory<UserSubWindow>::subscribe<TYPE>(realName);
+  saveAliasNames<TYPE>(realName);
+
+  // Make a record of each interface's categories.
+  const QStringList categories =
+      TYPE::categoryInfo().split(";", QString::SkipEmptyParts);
+  QSet<QString> result;
+  foreach (const QString category, categories) {
+    result.insert(category.trimmed());
+  }
+  m_categoryLookup[QString::fromStdString(realName)] = result;
+}
+
+/**
+ * Save the alias names of an interface
+ * @param realName :: The real name of the interface
+ */
+template <typename TYPE>
+void UserSubWindowFactoryImpl::saveAliasNames(const std::string &realName) {
+  std::set<std::string> aliases = TYPE::aliases();
+  for (const auto &alias_std_str : aliases) {
+    QString alias = QString::fromStdString(alias_std_str);
+    if (m_aliasLookup.contains(alias)) {
+      if (m_badAliases.contains(alias)) {
+        QList<std::string> names = m_badAliases.value(alias);
+        names.append(realName);
+        m_badAliases[alias] = names;
+      } else {
+        QList<std::string> names;
+        names.append(m_aliasLookup.value(alias));
+        names.append(realName);
+        m_badAliases.insert(alias, names);
+      }
+      continue;
+    }
+    m_aliasLookup.insert(alias, realName);
+  }
+}
+
+/**
+ * The keys associated with UserSubWindow classes
+ * @returns A QStringList containing the keys from the UserSubWindowFactory that
+ * refer to UserSubWindow classes
+ */
+QStringList UserSubWindowFactoryImpl::getUserSubWindowKeys() const {
+  QStringList key_list;
+  const auto keys = getKeys();
+  for (const auto &key : keys) {
+    key_list.append(QString::fromStdString(key));
+  }
+  return key_list;
 }
